requirements.txt
pillow>=9.0.0
numpy>=1.20.0
scipy>=1.6.0
pysstv>=0.9.6
simpleaudio>=1.0.4 